# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Callception API",
    "description": "AI-powered call management, appointment tracking, and customer service automation platform. Provides voice AI pipeline, CRM, ticketing, billing, and multi-tenant management capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "Callception",
      "url": "https://callception.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://callception.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "TenantHeader": []
    }
  ],
...
  "paths": {
...
    "/api/voice/session": {
      "get": {
        "tags": ["Voice Pipeline"],
        "summary": "Session status, personas, or context",
        "parameters": [
          { "name": "action", "in": "query", "schema": { "type": "string", "enum": ["status", "personas", "context"], "default": "status" } },
          { "name": "session_id", "in": "query", "schema": { "type": "string" }, "description": "Required for action=context" }
        ],
        "responses": { "200": { "description": "Session info or persona list" } }
      },
      "post": {
        "tags": ["Voice Pipeline"],
        "summary": "Save session or proxy inference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["save", "infer"], "default": "save" },
                  "sessionId": { "type": "string" },
                  "customerId": { "type": "string" },
                  "customerPhone": { "type": "string" },
                  "customerName": { "type": "string" },
                  "transcript": { "type": "array", "items": { "type": "object" } },
                  "duration": { "type": "number" },
                  "persona": { "type": "string" },
                  "text": { "type": "string", "description": "For action=infer" },
                  "language": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Session saved or inference result" }
        }
      }
    },
    "/api/voice/connect": {
      "get": {
        "tags": ["Voice Pipeline"],
        "summary": "WebSocket connection info",
        "description": "Returns WebSocket URL and audio format requirements for real-time voice.",
        "responses": { "200": { "description": "WebSocket connection details" } }
      },
      "post": {
        "tags": ["Voice Pipeline"],
        "summary": "Create authenticated session token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "persona": { "type": "string", "default": "default" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "expires_in": { "type": "integer" },
                    "websocket_url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing/webhook": {
      "post": {
        "tags": ["Billing"],
        "summary": "Lemon Squeezy webhook handler",
        "description": "Receives and processes subscription lifecycle events from Lemon Squeezy. Verifies HMAC-SHA256 signature.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "description": "Lemon Squeezy webhook payload" }
            }
          }
        },
        "responses": {
          "200": { "description": "Webhook processed" },
          "401": { "description": "Invalid signature" }
        }
      },
      "get": {
        "tags": ["Billing"],
        "summary": "Get subscription status",
        "responses": { "200": { "description": "Current subscription details" } }
      }
    },
...
    "/api/billing/alert-test": {
      "get": {
        "tags": ["Billing"],
        "summary": "Check alert channel status",
        "responses": { "200": { "description": "Alert channel configuration status" } }
      },
      "post": {
        "tags": ["Billing"],
        "summary": "Send test alert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["info", "warning", "critical", "emergency"], "default": "info" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Test alert dispatched" } }
      }
    },
...
    "/api/twilio/incoming": {
      "post": {
        "tags": ["Twilio"],
        "summary": "Incoming call webhook",
        "description": "Called by Twilio when a call arrives. Resolves tenant, validates subscription, returns TwiML with greeting and speech gather.",
        "security": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "type": "object", "description": "Twilio call event parameters" }
            }
          }
        },
        "responses": {
          "200": { "description": "TwiML response", "content": { "text/xml": {} } }
        }
      }
    },
    "/api/twilio/gather": {
      "post": {
        "tags": ["Twilio"],
        "summary": "Speech gather callback",
        "description": "Called by Twilio after speech recognition. Processes with LLM and returns TwiML with AI response.",
        "security": [],
        "parameters": [
          { "name": "tenantId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "callSid", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "TwiML with AI response", "content": { "text/xml": {} } }
        }
      }
    },
...
    "/api/twilio/test": {
      "get": {
        "tags": ["Twilio"],
        "summary": "Twilio configuration status",
        "responses": { "200": { "description": "Twilio and LLM provider status" } }
      },
      "post": {
        "tags": ["Twilio"],
        "summary": "Simulate call conversation",
        "description": "Tests the call flow without a real Twilio call.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": { "type": "string", "default": "Merhaba, randevu almak istiyorum" },
                  "language": { "type": "string", "default": "tr" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Simulated call result" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT",
        "description": "Firebase ID token for authenticated requests"
      },
      "TenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-tenant",
        "description": "Tenant ID header set by authentication middleware"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for automation/webhook endpoints"
      }
    },
...
  }
}